pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["tinumide <summittinuade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
sklearn = "^0.0.post7"
scikit-learn = "^1.3.0"
pandas = "^2.0.3"
pinecone-client = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

